Add input and echo
PAPA (Previously named Duke) now waits for input and will echo the
input.
Greeting messages and miscellaneous strings put in separate
Messages class.
